requirements.txt
numpy>=1.26
scipy>=1.11
pandas>=2.1
matplotlib>=3.8
torch>=2.2
PyYAML>=6.0
tqdm>=4.66
pytest>=8.0
